Added /todos endpoint and polished the codebase

+++ src/main/java/sokrisztian/todo/admin/persistance/repository/UserBasicRepository.java
@@ -3,5 +3,5 @@ package sokrisztian.todo.admin.persistance.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sokrisztian.todo.admin.persistance.domain.UserEntity;
 
-public interface UserBasicRepository extends JpaRepository<UserEntity, Long> {
+public interface UserBasicRepository extends JpaRepository<UserEntity, Integer> {
 }
